requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=5.4.0
tensorboard>=2.9.0
tqdm>=4.62.0
einops>=0.4.0
timm>=0.6.0
